requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-environ
python-telegram-bot
requests==2.32.3
python-dotenv==1.0.1